+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `rocksdb-storage` feature and associated items (See [PR #761](https://github.com/teloxide/teloxide/pull/761) for reasoning) [**BC**]
 
+## Changed
+
+- The following functions were made `const`:
+  - `Dialogue::{new, chat_id}`
+  - `Polling::builder`
+  - `StatefulListener::new_with_hints`
+  - `CommandDescriptions::new`
+  - `respond`
+
 ## 0.11.1 - 2022-10-31
 
 ### Added